Moved the refinements to different files and made a folder for them. Also deleted a redundant line in batchifier spec
