+++ lib/wor/batchifier/merge_params/object.rb
@@ -1,0 +1,15 @@
+module Wor
+  module Batchifier
+    module MergeParams
+      refine Object do
+        def merge_method
+          raise Wor::Batchifier::Exceptions::InvalidStrategyType
+        end
+
+        def merge_base_case
+          raise Wor::Batchifier::Exceptions::InvalidStrategyType
+        end
+      end
+    end
+  end
+end
