+++ lib/wor/batchifier/merge_params/proc.rb
@@ -1,0 +1,15 @@
+module Wor
+  module Batchifier
+    module MergeParams
+      refine Proc do
+        def merge_method
+          self
+        end
+
+        def merge_base_case
+          {}
+        end
+      end
+    end
+  end
+end
